move DocumentMetadataDeletionTestConfig
